add routes and finish day04
