+++ mobile/package.json
@@ -8,13 +8,22 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@expo-google-fonts/nunito": "^0.1.0",
+    "@react-native-community/masked-view": "0.1.10",
+    "@react-navigation/native": "^5.7.6",
+    "@react-navigation/stack": "^5.9.3",
     "expo": "~39.0.2",
+    "expo-font": "~8.3.0",
     "expo-status-bar": "~1.0.2",
     "maps": "^0.3.3",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "^0.63.3",
+    "react-native-gesture-handler": "~1.7.0",
     "react-native-maps": "0.27.1",
+    "react-native-reanimated": "~1.13.0",
+    "react-native-safe-area-context": "3.1.4",
+    "react-native-screens": "~2.10.1",
     "react-native-web": "~0.13.12"
   },
   "devDependencies": {
